copiloto-function/requirements.txt
azure-ai-agents==1.0.2
azure-ai-projects==1.0.0b12
azure-core==1.35.0
azure-identity>=1.15.0
azure-storage-blob>=12.20.0
azure-monitor-query==1.2.0
azure-monitor-opentelemetry>=1.8.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
idna==3.10
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
pycparser==2.22
PyJWT==2.10.1
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
azure-functions==1.20.0
azure-mgmt-compute

# Management plane SDKs
azure-mgmt-resource>=23.0.0,<24.0.0
azure-mgmt-web>=9.0.0
azure-mgmt-storage>=21.0.0,<22.0.0
azure-mgmt-monitor>=6.0.0,<7.0.0
azure-mgmt-authorization>=2.0.0,<3.0.0
azure-mgmt-compute
azure-mgmt-network

# Cosmos DB support
azure-cosmos>=4.2.0

# Azure AI Search support
azure-search-documents>=11.4.0

psutil==5.9.8

# Added per request
redis[hiredis,json]
